moving _common_ui_targets into _ui_tester_register
