+++ traitsui/testing/tester/_ui_tester_register/wx/_traitsui/check_list_editor.py
@@ -12,7 +12,7 @@ import wx
 
 from traitsui.wx.check_list_editor import CustomEditor
 from traitsui.testing.tester import command, locator
-from traitsui.testing.tester._common_ui_targets import _BaseSourceWithLocation
+from traitsui.testing.tester._ui_tester_register._common_ui_targets import _BaseSourceWithLocation
 from traitsui.testing.tester.editors._layout import column_major_to_row_major
 from traitsui.testing.tester._ui_tester_register.wx import _interaction_helpers
